requirements.txt
asgiref==3.9.2
certifi==2025.8.3
charset-normalizer==3.4.3
contourpy==1.3.3
cycler==0.12.1
Django==5.2.6
djangorestframework==3.16.1
fonttools==4.60.0
idna==3.10
kiwisolver==1.4.9
matplotlib==3.10.6
narwhals==2.5.0
numpy==2.3.3
packaging==25.0
pandas==2.3.2
pillow==11.3.0
plotly==6.3.0
psycopg2-binary==2.9.10
pyparsing==3.2.5
python-dateutil==2.9.0.post0
pytz==2025.2
requests==2.32.5
six==1.17.0
sqlparse==0.5.3
tzdata==2025.2
urllib3==2.5.0
